requirements.txt
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-text-splitters>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
